move config classes to subdirectory
